requirements.txt
asgiref==3.2.7
Django==3.0.6
django-environ==0.4.5
django-rest-framework==0.1.0
djangorestframework==3.11.0
pytz==2020.1
sqlparse==0.3.1
